refactor: modularize release gate flows and refresh pilot evidence
