+++ docs/pilot-evidence/index.csv
@@ -25,3 +25,4 @@ pilot-entry-20260316T114245Z-pilot-ks-rehearsal-2026-03-15,pilot-ks-rehearsal-20
 pilot-entry-20260316T120545Z-pilot-ks-rehearsal-2026-03-15,pilot-ks-rehearsal-2026-03-15,production,all,2026-03-16T12:05:45.138Z,GO,docs/release-gates/2026-03-16_production_dpl_5R7nc92c58ufkUnQeBmXFUAgbhu1.md,docs/pilot/PILOT_EVIDENCE_INDEX_pilot-ks-rehearsal-2026-03-15.md,
 pilot-entry-20260316T150046Z-pilot-ks-rehearsal-2026-03-15,pilot-ks-rehearsal-2026-03-15,production,all,2026-03-16T15:00:46.723Z,GO,docs/release-gates/2026-03-16_production_dpl_5R7nc92c58ufkUnQeBmXFUAgbhu1.md,docs/pilot/PILOT_EVIDENCE_INDEX_pilot-ks-rehearsal-2026-03-15.md,
 pilot-entry-20260316T204146Z-pilot-ks-7d-rehearsal-2026-03-16,pilot-ks-7d-rehearsal-2026-03-16,production,all,2026-03-16T20:41:46.095Z,GO,docs/release-gates/2026-03-16_production_dpl_5R7nc92c58ufkUnQeBmXFUAgbhu1.md,docs/pilot/PILOT_EVIDENCE_INDEX_pilot-ks-7d-rehearsal-2026-03-16.md,
+pilot-entry-20260318T200806Z-pilot-ks-process-proof-2026-03-18,pilot-ks-process-proof-2026-03-18,production,all,2026-03-18T20:08:06.833Z,NO-GO,docs/release-gates/2026-03-18_production_dpl_rKYXk6CrZ3p2KNPDidHiruc96WA7.md,docs/pilot/PILOT_EVIDENCE_INDEX_pilot-ks-process-proof-2026-03-18.md,
