+++ docs/pilot/live-data/pilot-ks-live-2026-03-18_day-2_claim-timeline-export.csv
@@ -1,7 +1,15 @@
 id,tenant_id,branch_id,member_id,claim_created_at,submitted_at,current_status,claim_id_history,to_status,is_public,history_created_at
-5e5868ab-62c5-456e-9263-b98ba12fc2b3,tenant_ks,KS,golden_ks_a_member_1,2026-03-19T09:00:00.000Z,2026-03-19T09:00:00.000Z,submitted,,,,
-442953f9-a4c9-4296-a76c-6671d5b59b76,tenant_ks,KS,golden_ks_a_member_1,2026-03-19T09:00:00.000Z,2026-03-19T09:00:00.000Z,submitted,,,,
-56d50e6e-1af7-493e-9f62-22d8bf448b19,tenant_ks,KS,golden_ks_a_member_1,2026-03-19T09:00:00.000Z,2026-03-19T09:00:00.000Z,submitted,,,,
-f74ef418-16d5-49ab-8475-21954538c9a4,tenant_ks,KS,golden_ks_a_member_1,2026-03-19T09:00:00.000Z,2026-03-19T09:00:00.000Z,submitted,,,,
-99020164-f685-4a6f-9a95-f2b2bdc64eb8,tenant_ks,KS,golden_ks_a_member_1,2026-03-19T09:00:00.000Z,2026-03-19T09:00:00.000Z,verification,,,,
-aab5c2a6-d85e-45b0-9a24-738da9d4e536,tenant_ks,KS,golden_ks_a_member_1,2026-03-19T09:00:00.000Z,2026-03-19T09:00:00.000Z,verification,,,,
+062e3b45-e42b-46eb-9ca2-f704a5eae18e,tenant_ks,,golden_ks_a_member_1,2026-03-19 09:00:00,2026-03-19 09:00:00,submitted,062e3b45-e42b-46eb-9ca2-f704a5eae18e,submitted,t,2026-03-19 09:00:00
+0e52bad3-0ebf-44dc-bde5-8783681a6c3e,tenant_ks,,golden_ks_a_member_1,2026-03-19 09:00:00,2026-03-19 09:00:00,verification,0e52bad3-0ebf-44dc-bde5-8783681a6c3e,verification,t,2026-03-19 09:00:00
+26d30168-3e2d-42f7-94da-800a6ec82571,tenant_ks,,golden_ks_a_member_1,2026-03-19 09:00:00,2026-03-19 09:00:00,verification,26d30168-3e2d-42f7-94da-800a6ec82571,verification,t,2026-03-19 09:00:00
+3d740446-cf8b-4c9c-9620-73e020b79d90,tenant_ks,,golden_ks_a_member_1,2026-03-19 09:00:00,2026-03-19 09:00:00,submitted,3d740446-cf8b-4c9c-9620-73e020b79d90,submitted,t,2026-03-19 09:00:00
+442953f9-a4c9-4296-a76c-6671d5b59b76,tenant_ks,,golden_ks_a_member_1,2026-03-19 09:00:00,2026-03-19 09:00:00,submitted,442953f9-a4c9-4296-a76c-6671d5b59b76,submitted,t,2026-03-19 09:00:00
+56d50e6e-1af7-493e-9f62-22d8bf448b19,tenant_ks,,golden_ks_a_member_1,2026-03-19 09:00:00,2026-03-19 09:00:00,submitted,56d50e6e-1af7-493e-9f62-22d8bf448b19,submitted,t,2026-03-19 09:00:00
+58ad97a3-8cc3-49a5-b7bb-0b24ede63f18,tenant_ks,,golden_ks_a_member_1,2026-03-19 09:00:00,2026-03-19 09:00:00,submitted,58ad97a3-8cc3-49a5-b7bb-0b24ede63f18,submitted,t,2026-03-19 09:00:00
+5b57ca89-d6cf-4ae4-b187-e542b1990066,tenant_ks,,golden_ks_a_member_1,2026-03-19 09:00:00,2026-03-19 09:00:00,submitted,5b57ca89-d6cf-4ae4-b187-e542b1990066,submitted,t,2026-03-19 09:00:00
+5e5868ab-62c5-456e-9263-b98ba12fc2b3,tenant_ks,,golden_ks_a_member_1,2026-03-19 09:00:00,2026-03-19 09:00:00,submitted,5e5868ab-62c5-456e-9263-b98ba12fc2b3,submitted,t,2026-03-19 09:00:00
+7c70080e-80e2-4cfe-a2bc-ac81f33055fa,tenant_ks,,golden_ks_a_member_1,2026-03-19 09:00:00,2026-03-19 09:00:00,submitted,7c70080e-80e2-4cfe-a2bc-ac81f33055fa,submitted,t,2026-03-19 09:00:00
+99020164-f685-4a6f-9a95-f2b2bdc64eb8,tenant_ks,,golden_ks_a_member_1,2026-03-19 09:00:00,2026-03-19 09:00:00,verification,99020164-f685-4a6f-9a95-f2b2bdc64eb8,verification,t,2026-03-19 09:00:00
+9f439933-5cac-4ac1-9818-17842edd36cb,tenant_ks,,golden_ks_a_member_1,2026-03-19 09:00:00,2026-03-19 09:00:00,submitted,9f439933-5cac-4ac1-9818-17842edd36cb,submitted,t,2026-03-19 09:00:00
+aab5c2a6-d85e-45b0-9a24-738da9d4e536,tenant_ks,,golden_ks_a_member_1,2026-03-19 09:00:00,2026-03-19 09:00:00,verification,aab5c2a6-d85e-45b0-9a24-738da9d4e536,verification,t,2026-03-19 09:00:00
+f74ef418-16d5-49ab-8475-21954538c9a4,tenant_ks,,golden_ks_a_member_1,2026-03-19 09:00:00,2026-03-19 09:00:00,submitted,f74ef418-16d5-49ab-8475-21954538c9a4,submitted,t,2026-03-19 09:00:00
