+++ docs/pilot/live-data/pilot-ks-live-2026-03-18_day-7_claim-timeline-export.csv
@@ -1,0 +1,1 @@
+id,tenant_id,branch_id,member_id,claim_created_at,submitted_at,current_status,claim_id_history,to_status,is_public,history_created_at
