+++ scripts/release-gate/product-checks.ts
@@ -1,0 +1,648 @@
+const fs = require('node:fs');
+const os = require('node:os');
+const path = require('node:path');
+
+const { MARKERS, ROUTES, SELECTORS, TIMEOUTS } = require('./config.ts');
+const {
+  buildRoute,
+  buildRouteAllowingLocalePath,
+  envFlag,
+  markerSnapshot,
+  sleep,
+} = require('./shared.ts');
+const { seedCookieConsentState } = require('./scenario-visits.ts');
+const { collectStaffClaimDetailUrls } = require('./staff-claim-driver.ts');
+const { gotoWithSessionRetry, loginWithRunContext } = require('./session-navigation.ts');
+
+const MEMBER_DOCUMENTS_CLAIM_CARD_SELECTOR = '[data-testid^="member-documents-claim-"]';
+const MEMBER_DOCUMENTS_CREATE_CLAIM_LINK_NAME = /create claim|krijo kërkesë/i;
+const DEFAULT_P13_STAFF_CLAIM_ROUTE = '/staff/claims/golden_ks_a_claim_05';
+const ACTIONABLE_CLAIM_STATUS_LABELS = new Set([
+  'Submitted',
+  'Verification',
+  'Evaluation',
+  'Negotiation',
+  'Court',
+]);
+const INFRA_NETWORK_ERROR_PATTERNS = [
+  /ECONNREFUSED/i,
+  /ETIMEDOUT/i,
+  /Timeout \d+ms exceeded/i,
+  /ENOTFOUND/i,
+  /EAI_AGAIN/i,
+  /ECONNRESET/i,
+  /socket hang up/i,
+  /AUTH_LOGIN_NETWORK_ERROR/i,
+];
+
+function compactErrorMessage(raw, maxLength = 420) {
+  return String(raw || '')
+    .replace(/\s+/g, ' ')
+    .trim()
+    .slice(0, maxLength);
+}
+
+function classifyInfraNetworkFailure(raw) {
+  const message = compactErrorMessage(raw, 650);
+  if (!message) return null;
+  const matched = INFRA_NETWORK_ERROR_PATTERNS.some(pattern => pattern.test(message));
+  return matched ? message : null;
+}
+
+function selectAlternativeActionableStatus(currentLabel, optionLabels) {
+  const normalizedCurrent = String(currentLabel || '')
+    .trim()
+    .toLowerCase();
+  for (const optionLabel of optionLabels) {
+    const normalizedOption = String(optionLabel || '').trim();
+    if (!normalizedOption || normalizedOption.toLowerCase() === normalizedCurrent) continue;
+    if (ACTIONABLE_CLAIM_STATUS_LABELS.has(normalizedOption)) {
+      return normalizedOption;
+    }
+  }
+  return null;
+}
+
+function trimTrailingSlashes(pathname) {
+  let end = pathname.length;
+  while (end > 1 && pathname.charCodeAt(end - 1) === 47) {
+    end -= 1;
+  }
+  return pathname.slice(0, end);
+}
+
+function resolveConfiguredStaffClaimDetailUrl(runCtx) {
+  const configured = String(process.env.STAFF_CLAIM_URL || '').trim();
+  if (!configured) {
+    return {
+      reason: 'missing',
+      source: 'missing',
+      url: null,
+    };
+  }
+
+  const targetUrl = /^https?:\/\//i.test(configured)
+    ? configured
+    : buildRouteAllowingLocalePath(runCtx.baseUrl, runCtx.locale, configured);
+
+  try {
+    const resolvedTarget = new URL(targetUrl);
+    const resolvedList = new URL(buildRoute(runCtx.baseUrl, runCtx.locale, ROUTES.staffClaimsList));
+    const normalizePath = pathname => trimTrailingSlashes(pathname) || '/';
+    if (normalizePath(resolvedTarget.pathname) === normalizePath(resolvedList.pathname)) {
+      return {
+        reason: 'list-url',
+        source: 'ignored-list',
+        url: null,
+      };
+    }
+  } catch {
+    return {
+      reason: 'invalid-url',
+      source: 'invalid',
+      url: null,
+    };
+  }
+
+  return {
+    reason: null,
+    source: 'env',
+    url: targetUrl,
+  };
+}
+
+async function runP11AndP12(browser, runCtx, deps) {
+  const { checkResult } = deps;
+  const evidenceP11 = [];
+  const signaturesP11 = [];
+  const evidenceP12 = [];
+  const signaturesP12 = [];
+
+  const now = Date.now();
+  const uploadName = `gate-upload-${now}.pdf`;
+  const uploadPath = path.join(os.tmpdir(), uploadName);
+  const uploadPdf = `%PDF-1.4
+1 0 obj
+<< /Type /Catalog /Pages 2 0 R >>
+endobj
+2 0 obj
+<< /Type /Pages /Kids [3 0 R] /Count 1 >>
+endobj
+3 0 obj
+<< /Type /Page /Parent 2 0 R /MediaBox [0 0 200 200] /Contents 4 0 R >>
+endobj
+4 0 obj
+<< /Length 44 >>
+stream
+BT /F1 12 Tf 20 100 Td (release-gate) Tj ET
+endstream
+endobj
+trailer
+<< /Root 1 0 R >>
+%%EOF
+`;
+  fs.writeFileSync(uploadPath, uploadPdf, 'utf8');
+
+  const context = await browser.newContext({ acceptDownloads: true });
+  const page = await context.newPage();
+  await seedCookieConsentState({ context, page, baseUrl: runCtx.baseUrl });
+  const clientSignals = [];
+  const pushClientSignal = signal => {
+    if (clientSignals.length < 10 && signal) {
+      clientSignals.push(String(signal).slice(0, 400));
+    }
+  };
+  let signedUploadStatuses = [];
+  let documentsDownloadStatuses = [];
+  let persistenceAfterRefresh = false;
+  let persistenceAfterRelogin = false;
+  const requireMemberUpload = envFlag('RELEASE_GATE_REQUIRE_MEMBER_UPLOAD', true);
+  const waitForUploadedFileVisible = async (targetPage, options = {}) => {
+    const timeoutMs = Number(options.timeoutMs ?? TIMEOUTS.upload);
+    const reloadBetweenAttempts = options.reloadBetweenAttempts === true;
+    const startedAt = Date.now();
+
+    while (Date.now() - startedAt < timeoutMs) {
+      const isVisible = await targetPage
+        .getByText(uploadName)
+        .first()
+        .isVisible({ timeout: TIMEOUTS.quickMarker })
+        .catch(() => false);
+      if (isVisible) {
+        return true;
+      }
+
+      if (reloadBetweenAttempts) {
+        await targetPage
+          .reload({ waitUntil: 'domcontentloaded', timeout: TIMEOUTS.nav })
+          .catch(() => {});
+      } else {
+        await targetPage.waitForTimeout(450);
+      }
+    }
+
+    return false;
+  };
+
+  page.on('response', response => {
+    const url = response.url();
+    if (url.includes('/storage/v1/object/upload/sign/')) {
+      signedUploadStatuses.push(`${response.status()}@${url}`);
+    }
+    if (url.includes('/api/documents/') && url.includes('/download')) {
+      documentsDownloadStatuses.push(`${response.status()}@${url}`);
+    }
+  });
+  page.on('console', msg => {
+    const text = msg.text();
+    const type = msg.type();
+    if (
+      type === 'error' ||
+      /upload flow error|storage upload unavailable|failed to generate upload url|mime type|supabase/i.test(
+        text
+      )
+    ) {
+      pushClientSignal(`console.${type}: ${text}`);
+    }
+  });
+  page.on('pageerror', error => {
+    pushClientSignal(`pageerror: ${String(error.message || error)}`);
+  });
+
+  try {
+    await loginWithRunContext(page, runCtx, 'member');
+
+    const docsUrl = buildRoute(runCtx.baseUrl, runCtx.locale, ROUTES.memberDocuments);
+    await page.goto(docsUrl, { waitUntil: 'domcontentloaded', timeout: TIMEOUTS.nav });
+    const claimCards = page.locator(MEMBER_DOCUMENTS_CLAIM_CARD_SELECTOR);
+    const claimCardCount = await claimCards.count();
+    evidenceP11.push(`member documents claim card count=${claimCardCount}`);
+    if (claimCardCount === 0) {
+      const createClaimVisible = await page
+        .getByRole('link', { name: MEMBER_DOCUMENTS_CREATE_CLAIM_LINK_NAME })
+        .isVisible({ timeout: TIMEOUTS.marker })
+        .catch(() => false);
+      evidenceP11.push(`member documents empty-state create claim visible=${createClaimVisible}`);
+      if (!createClaimVisible) {
+        signaturesP11.push('P1.1_MEMBER_DOCUMENTS_EMPTY_STATE_INVALID');
+      }
+      evidenceP12.push('skip_reason=valid_member_empty_state_no_claim_cards');
+      return [
+        checkResult('P1.1', signaturesP11.length ? 'FAIL' : 'PASS', evidenceP11, signaturesP11),
+        checkResult('P1.2', 'SKIPPED', evidenceP12, ['P1.2_SKIPPED_VALID_MEMBER_EMPTY_STATE']),
+      ];
+    }
+
+    const uploadButtons = page.locator(SELECTORS.memberDocumentsUploadButtons);
+    const uploadButtonsCount = await uploadButtons.count();
+    evidenceP11.push(`member documents upload button count=${uploadButtonsCount}`);
+    if (uploadButtonsCount === 0) {
+      if (requireMemberUpload) {
+        signaturesP11.push(
+          'P1.1_MISCONFIG_MEMBER_UPLOAD_PRECONDITION_NOT_MET reason=no_upload_entry_buttons'
+        );
+        return [
+          checkResult('P1.1', 'FAIL', evidenceP11, signaturesP11),
+          checkResult('P1.2', 'SKIPPED', evidenceP12, [
+            'P1.2_SKIPPED_DEPENDENCY_P1.1_MEMBER_UPLOAD_PRECONDITION',
+          ]),
+        ];
+      }
+      return [
+        checkResult('P1.1', 'SKIPPED', evidenceP11, [
+          'P1.1_SKIPPED_MEMBER_UPLOAD_PRECONDITION_NOT_MET',
+        ]),
+        checkResult('P1.2', 'SKIPPED', evidenceP12, [
+          'P1.2_SKIPPED_DEPENDENCY_P1.1_MEMBER_UPLOAD_PRECONDITION',
+        ]),
+      ];
+    }
+
+    const uploadButton = uploadButtons.first();
+    let uploadDialogOpened = false;
+    for (let attempt = 0; attempt < 3; attempt += 1) {
+      await uploadButton.scrollIntoViewIfNeeded().catch(() => {});
+      await uploadButton.click({ timeout: TIMEOUTS.action }).catch(() => {});
+      uploadDialogOpened = await page
+        .locator(SELECTORS.fileInput)
+        .isVisible({ timeout: TIMEOUTS.action })
+        .catch(() => false);
+      if (uploadDialogOpened) break;
+      await sleep(400);
+    }
+    evidenceP11.push(`upload dialog opened=${uploadDialogOpened}`);
+    if (!uploadDialogOpened) {
+      throw new Error('UPLOAD_DIALOG_NOT_OPEN');
+    }
+    await page.setInputFiles(SELECTORS.fileInput, uploadPath);
+    await page.getByRole('button', { name: SELECTORS.uploadButtonName }).click();
+    await page.waitForTimeout(1200);
+    const listedAfterSubmit = await waitForUploadedFileVisible(page, {
+      timeoutMs: TIMEOUTS.upload,
+      reloadBetweenAttempts: false,
+    });
+    evidenceP11.push(`upload file listed after submit=${listedAfterSubmit}`);
+
+    persistenceAfterRefresh = await waitForUploadedFileVisible(page, {
+      timeoutMs: TIMEOUTS.upload,
+      reloadBetweenAttempts: true,
+    });
+    evidenceP11.push(`after hard refresh listed=${persistenceAfterRefresh}`);
+
+    await context.close();
+
+    const reloginContext = await browser.newContext({ acceptDownloads: true });
+    const reloginPage = await reloginContext.newPage();
+    await seedCookieConsentState({
+      context: reloginContext,
+      page: reloginPage,
+      baseUrl: runCtx.baseUrl,
+    });
+    reloginPage.on('response', response => {
+      const url = response.url();
+      if (url.includes('/api/documents/') && url.includes('/download')) {
+        documentsDownloadStatuses.push(`${response.status()}@${url}`);
+      }
+    });
+
+    await loginWithRunContext(reloginPage, runCtx, 'member');
+    await reloginPage.goto(docsUrl, { waitUntil: 'domcontentloaded', timeout: TIMEOUTS.nav });
+    persistenceAfterRelogin = await waitForUploadedFileVisible(reloginPage, {
+      timeoutMs: TIMEOUTS.upload,
+      reloadBetweenAttempts: true,
+    });
+    evidenceP11.push(`after logout/login listed=${persistenceAfterRelogin}`);
+
+    const signed200 = signedUploadStatuses.some(entry => entry.startsWith('200@'));
+    evidenceP11.push(
+      `signed upload statuses: ${signedUploadStatuses.length ? signedUploadStatuses.join(' | ') : 'none captured'}`
+    );
+    if (clientSignals.length > 0) {
+      evidenceP11.push(`client signals: ${clientSignals.join(' || ')}`);
+    }
+    if (!persistenceAfterRefresh || !persistenceAfterRelogin) {
+      signaturesP11.push(
+        `P1.1_UPLOAD_PERSISTENCE_FAILED refresh=${persistenceAfterRefresh} relogin=${persistenceAfterRelogin} file=${uploadName}`
+      );
+    }
+    if (!signed200 && !(persistenceAfterRefresh && persistenceAfterRelogin)) {
+      signaturesP11.push(`P1.1_SIGNED_UPLOAD_NOT_CONFIRMED file=${uploadName}`);
+    }
+
+    try {
+      const fileLabel = reloginPage.getByText(uploadName).first();
+      await fileLabel.waitFor({ state: 'visible', timeout: TIMEOUTS.marker });
+      const fileRow = fileLabel
+        .locator('xpath=ancestor::div[contains(@class, "flex items-center justify-between")]')
+        .first();
+
+      await fileRow.getByRole('button', { name: SELECTORS.downloadButtonName }).first().click();
+      const has200DownloadResponse = await reloginPage
+        .waitForResponse(
+          response =>
+            response.url().includes('/api/documents/') &&
+            response.url().includes('/download') &&
+            response.status() === 200,
+          { timeout: TIMEOUTS.download }
+        )
+        .then(() => true)
+        .catch(() => false);
+      evidenceP12.push(`download response 200 observed=${has200DownloadResponse}`);
+      evidenceP12.push(
+        `download response statuses: ${documentsDownloadStatuses.length ? documentsDownloadStatuses.join(' | ') : 'none captured'}`
+      );
+
+      let inlineOpened = false;
+      const popupPromise = reloginPage
+        .waitForEvent('popup', { timeout: TIMEOUTS.download })
+        .then(async popup => {
+          await popup
+            .waitForLoadState('domcontentloaded', { timeout: TIMEOUTS.download })
+            .catch(() => {});
+          const notFound = await popup
+            .getByTestId(MARKERS.notFound)
+            .isVisible({ timeout: TIMEOUTS.quickMarker })
+            .catch(() => false);
+          inlineOpened = !notFound;
+          await popup.close().catch(() => {});
+        })
+        .catch(() => {});
+      await fileRow
+        .getByRole('button', { name: SELECTORS.inlineViewButtonName })
+        .first()
+        .click()
+        .catch(() => {});
+      await popupPromise;
+      evidenceP12.push(`inline/open action succeeded=${inlineOpened}`);
+
+      if (!has200DownloadResponse && !inlineOpened) {
+        signaturesP12.push(`P1.2_DOWNLOAD_FAILED file=${uploadName}`);
+      }
+    } catch (downloadError) {
+      signaturesP12.push(
+        `P1.2_EXCEPTION message=${String(downloadError.message || downloadError)}`
+      );
+    }
+
+    await reloginContext.close();
+  } catch (error) {
+    const rawMessage = String(error.message || error);
+    const infraMessage = classifyInfraNetworkFailure(rawMessage);
+    if (rawMessage.includes('NO_MEMBER_DOCUMENT_UPLOAD_BUTTONS')) {
+      signaturesP11.push(
+        'P1.1_MISCONFIG_MEMBER_UPLOAD_PRECONDITION_NOT_MET reason=no_upload_entry_buttons'
+      );
+      signaturesP12.push('P1.2_SKIPPED_DEPENDENCY_P1.1_MEMBER_UPLOAD_PRECONDITION');
+    } else if (infraMessage) {
+      signaturesP11.push(`P1.1_INFRA_NETWORK message=${infraMessage}`);
+      signaturesP12.push(`P1.2_INFRA_NETWORK_DEPENDENCY message=${infraMessage}`);
+    } else {
+      signaturesP11.push(`P1.1_EXCEPTION message=${compactErrorMessage(rawMessage, 650)}`);
+      signaturesP12.push('P1.2_DEPENDENCY_P1.1_EXCEPTION');
+    }
+    if (clientSignals.length > 0) {
+      evidenceP11.push(`client signals: ${clientSignals.join(' || ')}`);
+      signaturesP11.push(`P1.1_CLIENT_SIGNAL ${clientSignals[0]}`);
+    }
+    await context.close().catch(() => {});
+  } finally {
+    fs.rmSync(uploadPath, { force: true });
+  }
+
+  return [
+    checkResult('P1.1', signaturesP11.length ? 'FAIL' : 'PASS', evidenceP11, signaturesP11),
+    checkResult('P1.2', signaturesP12.length ? 'FAIL' : 'PASS', evidenceP12, signaturesP12),
+  ];
+}
+
+async function runP13(browser, runCtx, deps) {
+  const { checkResult } = deps;
+  const evidence = [];
+  const signatures = [];
+  const context = await browser.newContext();
+  const page = await context.newPage();
+  await seedCookieConsentState({ context, page, baseUrl: runCtx.baseUrl });
+  try {
+    await loginWithRunContext(page, runCtx, 'staff');
+
+    const claimsList = buildRoute(runCtx.baseUrl, runCtx.locale, ROUTES.staffClaimsList);
+    await page.goto(claimsList, { waitUntil: 'networkidle', timeout: TIMEOUTS.nav });
+    const staffReadyOnList = await page
+      .getByTestId(MARKERS.staff)
+      .isVisible({ timeout: TIMEOUTS.marker })
+      .catch(() => false);
+    evidence.push(`staff_claims_list_url=${claimsList}`);
+    evidence.push(`staff_page_ready_on_list=${staffReadyOnList}`);
+    if (!staffReadyOnList) {
+      signatures.push('P1.3_STAFF_PORTAL_NOT_READY');
+      return checkResult('P1.3', 'FAIL', evidence, signatures);
+    }
+
+    const requireClaimUrl =
+      String(process.env.RELEASE_GATE_REQUIRE_STAFF_CLAIM_URL || 'false').toLowerCase() === 'true';
+    const configuredClaim = resolveConfiguredStaffClaimDetailUrl(runCtx);
+    let detailUrl = null;
+
+    if (configuredClaim.url) {
+      detailUrl = configuredClaim.url;
+      evidence.push('claim_source=STAFF_CLAIM_URL');
+    } else {
+      if (requireClaimUrl) {
+        signatures.push(
+          configuredClaim.reason === 'list-url'
+            ? 'P1.3_MISCONFIG_STAFF_CLAIM_URL_DETAIL_REQUIRED'
+            : 'P1.3_MISCONFIG_STAFF_CLAIM_URL_REQUIRED'
+        );
+        return checkResult('P1.3', 'FAIL', evidence, signatures);
+      }
+      if (configuredClaim.reason === 'list-url') {
+        evidence.push('claim_source=STAFF_CLAIM_URL_IGNORED_LIST');
+      }
+
+      const deterministicDetailUrl = buildRoute(
+        runCtx.baseUrl,
+        runCtx.locale,
+        DEFAULT_P13_STAFF_CLAIM_ROUTE
+      );
+      await gotoWithSessionRetry({
+        page,
+        navigate: () =>
+          page.goto(deterministicDetailUrl, { waitUntil: 'networkidle', timeout: TIMEOUTS.nav }),
+        retryLogin: () => loginWithRunContext(page, runCtx, 'staff'),
+      });
+      const deterministicNotFound = await page
+        .getByTestId(MARKERS.notFound)
+        .isVisible({ timeout: TIMEOUTS.quickMarker })
+        .catch(() => false);
+      const deterministicReady = await page
+        .locator(SELECTORS.staffClaimDetailReady)
+        .isVisible({ timeout: TIMEOUTS.marker })
+        .catch(() => false);
+
+      if (!deterministicNotFound && deterministicReady) {
+        detailUrl = deterministicDetailUrl;
+        evidence.push('claim_source=deterministic_staff_claim');
+      } else {
+        evidence.push(
+          `deterministic_claim_ready=${deterministicReady} deterministic_claim_not_found=${deterministicNotFound}`
+        );
+
+        const fallback = await collectStaffClaimDetailUrls(page, runCtx);
+        evidence.push(`fallback_search_elapsed_ms=${fallback.elapsedMs}`);
+        evidence.push(`fallback_link_count=${fallback.urls.length}`);
+        if (fallback.urls.length === 0) {
+          signatures.push('P1.3_NO_TEST_DATA_STAFF_CLAIMS');
+          return checkResult('P1.3', 'SKIPPED', evidence, signatures);
+        }
+        detailUrl = fallback.urls[0];
+        evidence.push('claim_source=staff_claims_list');
+      }
+    }
+
+    await gotoWithSessionRetry({
+      page,
+      navigate: () => page.goto(detailUrl, { waitUntil: 'networkidle', timeout: TIMEOUTS.nav }),
+      retryLogin: () => loginWithRunContext(page, runCtx, 'staff'),
+    });
+
+    evidence.push(`claim_url=${detailUrl}`);
+    const notFoundOnDetail = await page
+      .getByTestId(MARKERS.notFound)
+      .isVisible({ timeout: TIMEOUTS.quickMarker })
+      .catch(() => false);
+    evidence.push(`detail_not_found=${notFoundOnDetail}`);
+    if (notFoundOnDetail) {
+      signatures.push('P1.3_MISCONFIG_STAFF_CLAIM_URL_UNREACHABLE');
+      return checkResult('P1.3', 'SKIPPED', evidence, signatures);
+    }
+
+    const staffReadyOnDetail = await page
+      .locator(SELECTORS.staffClaimDetailReady)
+      .isVisible({ timeout: TIMEOUTS.marker })
+      .catch(() => false);
+    evidence.push(`staff_page_ready_on_detail=${staffReadyOnDetail}`);
+    const detailReady = staffReadyOnDetail;
+    const actionPanelReady = await page
+      .locator(SELECTORS.staffClaimActionPanel)
+      .isVisible({ timeout: TIMEOUTS.marker })
+      .catch(() => false);
+    const claimSectionReady = await page
+      .locator(SELECTORS.staffClaimSection)
+      .isVisible({ timeout: TIMEOUTS.marker })
+      .catch(() => false);
+
+    evidence.push(
+      `detail_ready=${detailReady} action_panel_ready=${actionPanelReady} claim_section_ready=${claimSectionReady}`
+    );
+    if (!detailReady && !actionPanelReady && !claimSectionReady) {
+      signatures.push('P1.3_DETAIL_READY_MARKER_MISSING');
+      return checkResult('P1.3', 'FAIL', evidence, signatures);
+    }
+
+    const statusTrigger = page.locator(SELECTORS.claimStatusSelectTrigger);
+    const currentStatusLabel = (await statusTrigger.innerText()).trim();
+    await statusTrigger.click();
+    await page
+      .locator(SELECTORS.claimStatusListbox)
+      .waitFor({ state: 'visible', timeout: TIMEOUTS.action });
+    const options = page.locator(SELECTORS.claimStatusOption);
+    const optionCount = await options.count();
+    if (optionCount === 0) {
+      signatures.push('P1.3_STATUS_OPTIONS_MISSING');
+    } else {
+      const optionLabels = [];
+      for (let i = 0; i < optionCount; i += 1) {
+        optionLabels.push((await options.nth(i).innerText()).trim());
+      }
+      const selectedLabel = selectAlternativeActionableStatus(currentStatusLabel, optionLabels);
+      if (!selectedLabel) {
+        signatures.push(
+          `P1.3_NO_ACTIONABLE_STATUS_TRANSITION current_status=${currentStatusLabel} options=${optionLabels.join('|')}`
+        );
+      } else {
+        for (let i = 0; i < optionCount; i += 1) {
+          const candidate = optionLabels[i];
+          if (candidate === selectedLabel) {
+            await options.nth(i).click();
+            break;
+          }
+        }
+        evidence.push(`status_change=${currentStatusLabel} -> ${selectedLabel}`);
+        const noteValue = `gate-note-${Date.now()}`;
+        await page.fill(SELECTORS.claimStatusNote, noteValue);
+        await page.getByRole('button', { name: SELECTORS.claimUpdateButtonName }).click();
+        await page.waitForTimeout(1600);
+        await page.reload({ waitUntil: 'domcontentloaded', timeout: TIMEOUTS.nav });
+        await page.locator(SELECTORS.staffClaimDetailReady).waitFor({
+          state: 'visible',
+          timeout: TIMEOUTS.marker,
+        });
+
+        const noteVisible = await page
+          .locator(SELECTORS.staffClaimNote)
+          .getByText(noteValue)
+          .isVisible({ timeout: TIMEOUTS.marker })
+          .catch(() => false);
+        evidence.push(`note persisted=${noteVisible} note="${noteValue}"`);
+        if (!noteVisible) {
+          signatures.push(`P1.3_NOTE_NOT_PERSISTED note=${noteValue}`);
+        }
+
+        const persistedStatusLabel = (
+          await page.locator(SELECTORS.claimStatusSelectTrigger).innerText()
+        ).trim();
+        const statusPersisted = persistedStatusLabel
+          .toLowerCase()
+          .includes(selectedLabel.toLowerCase());
+        evidence.push(
+          `status persisted=${statusPersisted} expected="${selectedLabel}" actual="${persistedStatusLabel}"`
+        );
+        if (!statusPersisted) {
+          signatures.push(`P1.3_STATUS_NOT_PERSISTED expected="${selectedLabel}"`);
+        }
+      }
+    }
+  } catch (error) {
+    const markerState = await markerSnapshot(page).catch(() => ({
+      member: false,
+      agent: false,
+      staff: false,
+      admin: false,
+    }));
+    const screenshotPath = path.join(
+      os.tmpdir(),
+      `release-gate-p13-failure-${Date.now()}-${Math.random().toString(16).slice(2)}.png`
+    );
+    await page.screenshot({ path: screenshotPath, fullPage: true }).catch(() => {});
+    evidence.push(`failure_url=${page.url()}`);
+    evidence.push(
+      `failure_markers member=${markerState.member} agent=${markerState.agent} staff=${markerState.staff} admin=${markerState.admin}`
+    );
+    evidence.push(`failure_screenshot=${screenshotPath}`);
+    const notFoundVisible = await page
+      .getByTestId(MARKERS.notFound)
+      .isVisible({ timeout: TIMEOUTS.quickMarker })
+      .catch(() => false);
+    evidence.push(`failure_not_found=${notFoundVisible}`);
+    const rawMessage = String(error.message || error);
+    const infraMessage = classifyInfraNetworkFailure(rawMessage);
+    if (infraMessage) {
+      signatures.push(`P1.3_INFRA_NETWORK message=${infraMessage}`);
+    } else {
+      signatures.push(`P1.3_EXCEPTION message=${compactErrorMessage(rawMessage, 650)}`);
+    }
+  } finally {
+    await context.close().catch(() => {});
+  }
+
+  return checkResult('P1.3', signatures.length ? 'FAIL' : 'PASS', evidence, signatures);
+}
+
+module.exports = {
+  classifyInfraNetworkFailure,
+  compactErrorMessage,
+  resolveConfiguredStaffClaimDetailUrl,
+  runP11AndP12,
+  runP13,
+  selectAlternativeActionableStatus,
+};
